requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
torch==2.0.1
torchvision==0.15.2
